Move svg component to generic shared folder
