+++ app/components/svg.tsx
@@ -58,3 +58,22 @@ export function LinkedinSvg() {
     </svg>
   );
 }
+
+export function ExternalLinkSvg() {
+  return (
+    <svg
+      xmlns={xmlns}
+      fill={fill}
+      viewBox="0 0 36 36"
+      strokeWidth={strokeWidth}
+      stroke={stroke}
+      className={className}
+    >
+      <path
+        strokeLinecap={strokeLinecap}
+        strokeLinejoin={strokeLinejoin}
+        d="M4.5 19.5l15-15m0 0H8.25m11.25 0v11.25"
+      />
+    </svg>
+  );
+}
